refactor(#792): split runtime_tests.rs by responsibility (3175->621 fixtures hub + 7 concern files)
Grouped 131 tests by what they exercise: block-assembly, frame DSP, frame-buffer,
graph, effective-io, process, integration. Shared fixtures/structs stay in the base
runtime_tests.rs (used across concerns). cargo test -p engine runtime: 517 pass.
